separate filter (StatusAwareMetricsFilter)
